feat: IssueTable 구현
