+++ FE/src/components/Main/IssuesList/IssueTable.jsx
@@ -1,80 +1,33 @@
-import { useState } from "react";
-import styled from "styled-components";
-import { Dropdown } from "../../../icons/dropdown";
-import { Closed } from "../../../icons/closed";
 import { Open } from "../../../icons/open";
-import { Popup } from "./Popup";
-
-const assigneePopupItems = [{ id: "no_assignee", label: "담당자가 없는 이슈" }];
-
-const labelPopupItems = [{ id: "no_label", label: "레이블이 없는 이슈" }];
-
-const milestonesPopupItems = [
-  { id: "no_milestones", label: "마일스톤이 없는 이슈" },
-];
-
-const authorPopupItems = [{ id: "no_author", label: "작성자가 없는 이슈" }];
-
-const headerFilters = [
-  { id: "assignee", label: "담당자", items: assigneePopupItems },
-  { id: "label", label: "레이블", items: labelPopupItems },
-  { id: "milestones", label: "마일스톤", items: milestonesPopupItems },
-  { id: "author", label: "작성자", items: authorPopupItems },
-];
+import styled from "styled-components";
+import { IssueTableHeader } from "./IssueTableHeader";
+import { issues } from "../../../test.json"; // test data
 
 export function IssueTable() {
-  const [popupStates, setPopupStates] = useState({
-    assignee: false,
-    label: false,
-    milestones: false,
-    author: false,
-  });
-
-  const handleOpenContent = (id) => {
-    setPopupStates((prev) => ({ ...prev, [id]: !prev[id] }));
-  };
-
-  const handleCloseContent = (id) => {
-    setPopupStates((prev) => ({ ...prev, [id]: false }));
-  };
-
   return (
     <StyledDiv>
-      <Header>
-        <Left>
-          <input type="checkbox" />
-          <StyledIssueBtn>
-            <Open />
-            <div>열린 이슈( )</div>
-          </StyledIssueBtn>
-          <StyledIssueBtn>
-            <Closed />
-            <div>닫힌 이슈( )</div>
-          </StyledIssueBtn>
-        </Left>
-        <Right>
-          {headerFilters.map((popup) => (
-            <Container key={popup.id}>
-              <Input
-                id={popup.id}
-                type="checkbox"
-                onChange={() => handleOpenContent(popup.id)}
-              />
-              <Filter htmlFor={popup.id}>
-                <div>{popup.label}</div>
-                <Dropdown />
-              </Filter>
-              <StyledPopup
-                id={popup.label}
-                isopen={popupStates[popup.id]}
-                popupItems={popup.items}
-                onChange={() => handleCloseContent(popup.id)}
-              />
-            </Container>
-          ))}
-        </Right>
-      </Header>
-      <Issues></Issues>
+      <IssueTableHeader />
+      {issues
+        .slice()
+        .reverse()
+        .map((issue, index) => (
+          <Issues key={index}>
+            <Checkbox type="checkbox" />
+            <Content>
+              <div className="top">
+                <Open />
+                <div className="title">{issue.title}</div>
+                <div>{issue.labels}</div>
+              </div>
+              <div className="bottom">
+                <div>#{issue.issue_id}</div>
+                <div>이 이슈가 {issue.create_time}전, {issue.assginee}님에 의해작성되었습니다
+                </div>
+                <div>{issue.milestone}</div>
+              </div>
+            </Content>
+          </Issues>
+        ))}
     </StyledDiv>
   );
 }
@@ -87,62 +40,31 @@ const StyledDiv = styled.div`
   border-radius: 10px;
 `;
 
-const Header = styled.div`
-  height: 60px;
-  border-bottom: solid #dadbef;
-  display: flex;
-  justify-content: space-between;
-`;
-
-const Left = styled.div`
-  display: flex;
-  align-items: center;
-  width: 300px;
-  justify-content: space-between;
-  margin-left: 20px;
-`;
-
-const StyledIssueBtn = styled.div`
-  display: flex;
-  justify-content: space-around;
-  width: 110px;
-`;
-
-const Right = styled.div`
-  display: flex;
-  align-items: center;
-  width: 450px;
-  justify-content: space-between;
-  margin-right: 20px;
-  position: relative;
-`;
-
-const Container = styled.div`
-  display: flex;
-  position: relative;
-`;
-
-const StyledPopup = styled(Popup)`
-  right: 0;
-`;
-
-const Input = styled.input`
-  left: 0;
-  visibility: hidden;
-  position: absolute;
-`;
-
-const Filter = styled.label`
-  display: flex;
-  justify-content: space-between;
-  border-radius: 10px;
-  padding: 10px;
-  align-items: center;
-  &:hover {
-    cursor: pointer;
-  }
-`;
-
 const Issues = styled.div`
+  display: flex;
   height: 90px;
+  background-color: white;
+  z-index: -1;
+`;
+
+const Content = styled.div`
+  display: flex;
+  flex-direction: column;
+  justify-content: space-evenly;
+  margin-left: 35px;
+  div {
+    margin-right: 30px;
+  }
+  .title {
+    margin-left: 10px;
+  }
+  .top,
+  .bottom {
+    display: flex;
+`;
+
+const Checkbox = styled.input`
+  height: 50%;
+  padding: 20px;
+  margin-left: 25px;
 `;
